mapping_mediator/requirements/development.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.20
django-environ==0.10.0
django-filter==23.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
email-validator==2.0.0.post2
fhir.resources==7.0.2
idna==3.4
importlib-metadata==6.7.0
Markdown==3.4.4
pydantic==1.10.12
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.15.0
